Remove staking requirement
