--- backend/grant/contribution/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import views
