+++ frontend/client/modules/ccr/utils.ts
@@ -59,7 +59,5 @@ export function getCCRErrors(
     }
   }
 
-  console.log('errors', errors);
-
   return errors;
 }
